re-jigged order in attempt to fix midi ccs
Tried a few things to modify the sound.

The way I'm handling the sensors & mapping them to volume appears to create a baseline of noise on the vs1053 output (might need to check the scaling)

I need to find something different for the MIDI ccs to do in order to move forward; volume isn't really cutting it.